backend-app-api: move http router factory implementation into subfolder
